+++ lib/ws_client.rb
@@ -1,0 +1,11 @@
+require 'event_emitter'
+require 'websocket'
+require 'socket'
+require 'openssl'
+require 'uri'
+
+require 'ws_client/version'
+require 'ws_client/client'
+
+module WSClient
+end